+++ src/composables/use-paginated-api.ts
@@ -40,22 +40,7 @@ export default function usePaginatedBackend (
     }
 
     isLoading.value = true
-
-    let response
-
-    try {
-      response = (await api.get(endpoint, { params: queryParams })).data
-    } catch (e) {
-      const error = e as any
-
-      if (error.response?.status && typeof (window) !== 'undefined') {
-        const event = new Event(`Error ${error.response?.status}`, { bubbles: true })
-        window.dispatchEvent(event)
-      }
-
-      return
-    }
-
+    const response = (await api.get(endpoint, { params: queryParams })).data
     count.value = response[responseTotalKey]
     const responseItems = response[responseItemsKey].map((i: any) => itemConverter(i))
     items.value = scrollPagination ? [...items.value, ...responseItems] : responseItems